Add modified MuJoCo envs better adapted to reward modeling (#6)
* Add modified HalfCheetah env better adapted to IRL

We create a modified HalfCheetah environment with:
- observation with center of mass position: this is because the reward depends on the center of mass position, and it simplifies the problem for IRL algorithms to figure out the reward if it is observable;
- no termination conditions and fixed time horizon, since termination conditions can be problematic by default to IRL algorithms (https://openreview.net/pdf?id=Hk4fpoA5Km)

* Add docstrings and use gym.wrappers.TimeLimit

We also:
- use standard naming for environment;
- restore test_dummy for the moment.

* Use evaluate_policy from stable_baselines

* Add all MuJoCo envs with v3 available

* Reuse max_episode_steps from gym, mark slow tests

* Add docstrings, reorder imports

We also add stable-baselines to TESTS_REQUIRE.

* Move env registration to mujoco.py

setup.py currently imports src.benchmark_environments to get the
version, so src.benchmark_environments should not import gym, since it
might be unavailable when setup.py starts (in which case installation
then fails). Thus we move for now the environment registration to
mujoco.py.

* Make isort use trailing comma, other fixes

We also make the benchmark_environments.mujoco unused import
to be ignored, since it does the environment registration.
These implicit registrations are not that great though, maybe it would
make sense to have an explicit function that registers the environments.

* Make docstring span one line

* Apply black formatting

* Increase stable-baselines version.

For some reason, tensorflow is not being installed in circleci testing even though it is supposedly a dependency of stable-baselines. Maybe forcing a recent version of stable-baselines fixes it.

* Import gym.envs to fix pytype error

* Add explicit tensorflow dependency

For some reason tensorflow is missing in circleci, even though it is a
dependency of stable-baselines. This might remove errors, though it is
not clear to me why tensorflow is not being installed.

* Add simple unit tests for new environments.

We add the simple env tests from
https://github.com/HumanCompatibleAI/imitation/blob/19ebc7216a220ee8290074b2faa09996216dffdb/tests/test_envs.py

* Set expensive tests off by default.

Expensive tests can now be activated by running
pytest with the --expensive flag.

* Fix docstrings in test files.

* Refactor: move tests into helper methods in separate file

* Fix test coverage

* Remove dummy tests (no longer needed)

* Update src/benchmark_environments/testing/envs.py

Co-Authored-By: Steven H. Wang <wang.steven.h@gmail.com>

* Update src/benchmark_environments/testing/envs.py

Co-Authored-By: Steven H. Wang <wang.steven.h@gmail.com>

* Update tests/test_mujoco_rl.py

Co-Authored-By: Steven H. Wang <wang.steven.h@gmail.com>

Co-authored-by: pedrofreire <pedrofreirex@gmail.com>
Co-authored-by: Steven H. Wang <wang.steven.h@gmail.com>